Update KlipTok.fr.json
Before it said "wrote {0} {1}: " so i made it so it says "At {0},  {1} wrote: " 
Also I don't know if you are aware but now everything is formally formulated while the readme says to translate informally but in french if i am not mistaken they usually go for formal with such things.

PS: I am not really a native french speaker. I live in Belgium and am dutch speaking natively but i am in a french speaking scouting group. So i do kinda know what i am talking about but take my advice with a grain of salt.

PSS: Love the project goodluck with it.

My discord if you have questions and or want further help : JOJO#9492